Stream Combining OnDiskSA Parts to Reduce Memory Overhead
When SASI combines parts of an index during flush data was
unnecessarily copied into memory due to CombinedTerm and
TokenTreeBuilder's use of SortedMap's, which not only introduced
unnecessary copying but retained the data for much longer than
required. In other cases, the SortedMap's are necessary during index
building because they preserve the token order required for the
TokenTreeBuilder.  However, when combining SA parts, the data is
already pre-sorted, introducing the uncessarary overhead that
enventually lead to an OOM on a 32GB heap with a 23GB sstable (with
about 2:1 compression ratio) - see CASSANDRA-11383.

This patch addresses this issue by exploiting the pre-sorted data in
the SA parts.  It does this by using using a RangeIterator to provide
a sorted stream of the parts, and by introducing a new version of the
TokenTreeBuilder: StaticTokenTreeBuilder -- the original
TokenTreeBuilder has been renamed DynamicTokenTreeBuilder. The
StaticTokenTreeBuilder uses the RangeIterator, now provided by
CombinedTerm, to iterate the data twice. The first iteration builds
the tree with leaves that contain only enough information to build the
upper layers -- these leaves do not store more than their minimum and
maximum tokens plus their total size, which makes them
unwriteable. When the tree is written to disk the final layer is not
written. Its at this point the data is iterated once again to write
the leaves to disk. This (logarithmically) reduces copying of the
token values while building and writing upper layers of the tree,
removes the use of SortedMap when combining SAs, and relies on the
memory mapped SAs otherwise, greatly improving performance and no
longer causing OOMs on the previously mentioned configuration.

NOTE: the tests in TokenTreeTest that tested
(Dynamic)TokenTreeBuilder's in memory iterator have been removed
because the iterator no longer iterates the built tree, opting to
iterate the SortedMap instead, negating the benefit of the test.