Add data layer to TokenTree
* The data layer contains partition offsets associated, in almost all
cases, with one or more Clustering's representing the full primary key
of the indexed value. This layer is of variable length, while the
tree, or index layer, remains block size based.

* The tree is backwards compatible with the existing AB index version.
8 ununsed bytes in the root header are used to indicate the location
of the data layer. One ununsed bit in each entry header is used to
indicate if the entry is packed into the index layer or the data
layer. The bit is unset, as in older versions, if the data is in the
index layer.

* The data layer exists but isn't used by the SASI subsystem after
this commit.  That means indexes written after this commit still
require reading the entire partition. This work is preparation for the
changes necessary to read only parts of the partition as
CASSANDRA-11990 intended.

* The read of the data layer doesn't properly account for MappedBuffer
not allowing reads that cross a page boundary. Tests will be added to
illustrate this case before fixing. The prep work for this, such as merging
clusterings into exsting entriues, is done as part of this commit.